Instanciation d'un épigraphe à partir de l'ID.
Il est apparu que plusieurs classes avaient besoin de générer des
entités Épigraphes avec un ID. Il est donc commode de déléguer cette
tâche au SI avec une méthode statique.

Cette modification utilise un code déjà présent dans l'API.
On y répercute donc la modification.

+++ backend/src/main/java/fr/univtln/m1infodid/projet_s2/backend/server/Api.java
@@ -40,8 +40,8 @@ public class Api {
 			JsonNode annotations = rootNode.get("annotations");
 			for (JsonNode jsonPoints : annotations) {
 				for (JsonNode coordonner : jsonPoints) {
-					Integer y = (Integer.parseInt(String.valueOf(coordonner.toString().charAt(1))));
-					Integer x = (Integer.parseInt(String.valueOf(coordonner.toString().charAt(3))));
+					double y = (Double.parseDouble(String.valueOf(coordonner.toString().charAt(1))));
+					double x = (Double.parseDouble(String.valueOf(coordonner.toString().charAt(3))));
 					annotation.get().addPoints(x, y);
 				}
 			}
